⚙️ move the code to read the taxonomy out of the main file
