+++ src/NCBITaxonomy.jl
@@ -1,15 +1,15 @@
 module NCBITaxonomy
 using DataFrames
-using Arrow
+import Arrow
 using StringDistances
 using AbstractTrees
 
 # Point to where the taxonomy is located
-include("hydrate.jl")
-local_path = _local_archive_path()
+include("local_archive_path.jl")
+tables_path = _create_or_get_tables_path(_local_archive_path())
 
 function __init__()
-    name_date = mtime(joinpath(local_path, "tables", "names.arrow"))
+    name_date = mtime(joinpath(tables_path, "names.arrow"))
     over_30_days = time() - name_date >= 2.6e+6
     if over_30_days 
         @warn("Your local taxonomy version is over 30 days old, we recommend using `] build NCBITaxonomy` to get the most recent version.")
@@ -23,30 +23,9 @@ export NCBITaxon, NCBINameClass, IDNotFoundInBackbone
 include("exceptions.jl")
 export NameHasNoDirectMatch, NameHasMultipleMatches
 
-names_table = DataFrame(Arrow.Table(joinpath(taxpath, "tables", "names.arrow")))
-names_table.class = NCBINameClass.(names_table.class)
-names_table.lowercase = lowercase.(names_table.name)
-
-division_table = DataFrame(Arrow.Table(joinpath(taxpath, "tables", "division.arrow")))
-select!(division_table, Not(:comments))
-
-nodes_table = DataFrame(Arrow.Table(joinpath(taxpath, "tables", "nodes.arrow")))
-select!(nodes_table, Not(r"inherited_"))
-select!(nodes_table, Not(r"_code_id"))
-select!(nodes_table, Not(:genbank_hidden))
-select!(nodes_table, Not(:hidden_subtree))
-select!(nodes_table, Not(:comments))
-select!(nodes_table, Not(:embl))
-
-nodes_table = innerjoin(nodes_table, division_table; on = :division_id)
-select!(nodes_table, Not(:division_id))
-
-names_table = leftjoin(
-    names_table,
-    unique(select(nodes_table, [:tax_id, :rank, :parent_tax_id]));
-    on = :tax_id,
-)
-scinames_table = names_table[findall(names_table.class .== class_scientific_name), :]
+# We load the core file with all we need in it
+include("read_taxonomy.jl")
+taxonomy = read_taxonomy(tables_path)
 
 include("taxon.jl")
 export taxon, @ncbi_str
